seed_engine/pyproject.toml
[project]
name = "seed-engine"
version = "1.0.0"
description = "Autonomous task execution agent — core engine"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "CodeAgent Contributors" }]
dependencies = [
    "requests>=2.28.0",
]

[project.optional-dependencies]
yaml = ["pyyaml>=6.0"]
dev = ["pytest>=7.0", "ruff>=0.4.0", "bandit>=1.7.0"]

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "SIM"]
ignore = ["E501", "N818"]
